Use mediumint

+++ db/migrate/20221201144756_add_num_finished_discussions_to_user_track_mentorships.rb
@@ -1,7 +1,7 @@
 class AddNumFinishedDiscussionsToUserTrackMentorships < ActiveRecord::Migration[7.0]
   def change
     unless Rails.env.production?
-      add_column :user_track_mentorships, :num_finished_discussions, :integer, null: false, default: 0
+      add_column :user_track_mentorships, :num_finished_discussions, :mediumint, null: false, default: 0
     end
   end
 end
